pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "oneconnect-linux"
version = "0.1.0"
description = "Clavister NetWall OIDC + OpenConnect helper with reusable core and GTK UI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "aiohttp>=3.10",
  "PyJWT>=2.8",
]

[project.scripts]
oneconnect = "oneconnect_core.cli:main"
oneconnect-gui = "oneconnect_gui.app:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
